+++ kernel/build_kernel
@@ -17,7 +17,6 @@ $SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-z
 $SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-zone -mno-mmx -mno-sse -mno-sse2 -mgeneral-regs-only -c drivers/ahci.c -o ahci.o || exit
 $SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-zone -mno-mmx -mno-sse -mno-sse2 -mgeneral-regs-only -c drivers/timer.c -o timer.o || exit
 $SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-zone -mno-mmx -mno-sse -mno-sse2 -mgeneral-regs-only -c drivers/multitasking.c -o multitasking2.o || exit
-$SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-zone -mno-mmx -mno-sse -mno-sse2 -mgeneral-regs-only -c drivers/xhci.c -o xhci.o || exit
 $SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-zone -mno-mmx -mno-sse -mno-sse2 -mgeneral-regs-only -c drivers/rtl.c -o rtl.o || exit
 $SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-zone -mno-mmx -mno-sse -mno-sse2 -mgeneral-regs-only -c drivers/ethernet.c -o ethernet.o || exit
 $SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-zone -mno-mmx -mno-sse -mno-sse2 -c drivers/fs/fat.c -o fat.o || exit
@@ -30,7 +29,7 @@ $SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-z
 $SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-zone -mno-mmx -mno-sse -mno-sse2 -c drivers/exec/debugger.c -o debugger.o || exit
 
 echo "Linking O files..."
-$SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -T ./linker.ld kernel.o graphics.o memory.o paging.o psf.o gdt.o debugger.o gdtsetup.o idt.o pci.o isr.o tty.o program.o ports.o ahci.o module.o elf.o multitasking.o multitasking2.o device.o xhci.o mbr.o fat.o timer.o ethernet.o rtl.o -o ./../bin/sanderos/kernel.bin -nostdlib -lgcc || exit
+$SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -T ./linker.ld kernel.o graphics.o memory.o paging.o psf.o gdt.o debugger.o gdtsetup.o idt.o pci.o isr.o tty.o program.o ports.o ahci.o module.o elf.o multitasking.o multitasking2.o device.o mbr.o fat.o timer.o ethernet.o rtl.o -o ./../bin/sanderos/kernel.bin -nostdlib -lgcc || exit
 
 echo "Cleanup"
 rm *.o  || exit
